osdrv: weekly rls 2024.10.14
-cc51f3, Merge "[fix](vi): Remove mandatory settings
-97b7e8, Merge "[refactor]( cvi_comm_aio.h): Invalid interface cleaning & interface adding
-611ff3, Merge "[feat](base):Modification of uintxx_t/intxx_t Data Types." into v4.1.0
-7ec541, [fix](RGN):[V4.1.0]Add structure comments for cvi_comm_region.h
-46e95b, [feat](base):Modification of uintxx_t/intxx_t Data Types.
...
-5fbc48, [feat](vpss):update vpss proc

+++ osdrv/interdrv/include/common/uapi/linux/cvi_type.h
@@ -44,8 +44,8 @@ typedef double                  CVI_DOUBLE;
 typedef void                    CVI_VOID;
 typedef unsigned char           CVI_BOOL;
 
-typedef uint64_t                CVI_U64;
-typedef int64_t                 CVI_S64;
+typedef unsigned long long int  CVI_U64;
+typedef long long int           CVI_S64;
 
 typedef unsigned int            CVI_SIZE_T;
